+++ lib/command/browser.rb
@@ -36,6 +36,7 @@ module Command
         puts @opt.help
         return
       end
+      tagname_to_ids(argv)
       argv.each do |target|
         data = Downloader.get_data_by_target(target)
         unless data