+++ lib/command/inspect.rb
@@ -39,6 +39,7 @@ module Command
         end
         return
       end
+      tagname_to_ids(argv)
       argv.each_with_index do |target, i|
         Helper.print_horizontal_rule if i > 0
         data = Downloader.get_data_by_target(target)